chức năng màn hình 8,9

+++ app/src/main/java/com/example/clientuser/database/object/Rating.java
@@ -4,12 +4,12 @@ public class Rating {
     String id_Store;
     String id_User;
     String comment;
-    int rating;
+    float rating;
 
     public Rating() {
     }
 
-    public Rating(String id_Store, String id_User, String comment, int rating) {
+    public Rating(String id_Store, String id_User, String comment, float rating) {
         this.id_Store = id_Store;
         this.id_User = id_User;
         this.comment = comment;
@@ -40,11 +40,11 @@ public class Rating {
         this.comment = comment;
     }
 
-    public int getRating() {
+    public float getRating() {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(float rating) {
         this.rating = rating;
     }
 }
